TF03/backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
mysql-connector-python==8.0.33
